+++ app/src/main/java/rkr/simplekeyboard/inputmethod/keyboard/internal/MoreKeySpec.java
@@ -70,7 +70,7 @@ public final class MoreKeySpec {
         mIconId = KeySpecParser.getIconId(moreKeySpec);
     }
 
-    public Key buildKey(final int x, final int y, final int labelFlags,
+    public Key buildKey(final float x, final float y, final int labelFlags,
             final KeyboardParams params) {
         return new Key(mLabel, mIconId, mCode, mOutputText, null /* hintLabel */, labelFlags,
                 Key.BACKGROUND_TYPE_NORMAL, x, y, params.mDefaultKeyWidth, params.mDefaultRowHeight,
